+++ packages/frontend/src/pages/Reports/components/CustomReportDialog.tsx
@@ -10,8 +10,18 @@ import { Button } from '@/components/ui/Button';
 import { makeClient } from '../../../utils/axios';
 import { ReportManifestDto } from '../../../model/report/dto/report-manifest.dto';
 import { isResponseAxiosError } from '../../../model/api';
+import { recoilPersist } from 'recoil-persist';
+import { atom, useRecoilState } from 'recoil';
 
-interface ReportConfigDialogProps {
+const { persistAtom } = recoilPersist();
+
+export const CustomReportUrl = atom<string>({
+  key: 'CustomReportUrl',
+  default: undefined,
+  effects: [persistAtom],
+});
+
+interface CustomReportDialogProps {
   onClose(): void;
   onRun(url: string, manifest: ReportManifestDto): void;
 }
@@ -19,9 +29,12 @@ interface ReportConfigDialogProps {
 export const CustomReportDialog = ({
   onClose,
   onRun,
-}: ReportConfigDialogProps): JSX.Element => {
-  const [url, setUrl] = React.useState('');
+}: CustomReportDialogProps): JSX.Element => {
+  // Local state
   const [error, setError] = React.useState('');
+
+  // Global state
+  const [url, setUrl] = useRecoilState(CustomReportUrl);
 
   const handleUrlChange = (event): void => {
     setUrl(event.target.value);
@@ -100,7 +113,7 @@ export const CustomReportDialog = ({
                 icon={faFileDownload}
                 size="1x"
               />
-              Run report
+              Load report
             </Button>
           </div>
         </div>