+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- **Frontend**: Custom reporst can now be confihured using an interactive form. #1131
 - **Frontend**: New feature: Run custom reports from the reports page. #1050
 
 ### Fixed